VENTURELINK/main-project/flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
SQLAlchemy
psycopg2
scikit-learn  # For AI matchmaking
pandas  # For data manipulation
python-dotenv  # To manage environment variables
openai

